pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "maturin>=1.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cxa-crypto"
version = "1.0.0"
description = "Advanced, OpSec-conscious cryptographic toolkit for data protection, steganography, and secure key management"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "XvoidcrewX", email = "voidcrew@cock.li"}
]
keywords = [
    "cryptography", "security", "encryption", "steganography",
    "key-management", "privacy", "security-tools", "aes",
    "chacha20", "rsa", "ed25519"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Apple :: macOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Security :: Security Tools"
]
dependencies = [
    "cryptography>=41.0.0",
    "numpy>=1.21.0",
    "Pillow>=9.0.0",
    "psutil>=5.8.0",
    "keyring>=23.0.0",
    "PyWavelets>=1.1.0",
    "zstandard>=0.15.0",
    " reedsolo>=1.7.0",
    "secretsharing>=0.2.7",
    "cffi>=1.15.0",
    "tkintermap-widget>=0.0.9; sys.platform == 'win32'"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-benchmark>=3.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
    "pre-commit>=3.0.0"
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0"
]
gui = [
    "customtkinter>=5.2.0"
]
all = [
    "cxa-crypto[dev]",
    "cxa-crypto[docs]",
    "cxa-crypto[gui]"
]

[project.urls]
Homepage = "https://github.com/XvoidcrewX/CXA"
Repository = "https://github.com/XvoidcrewX/CXA"
Issues = "https://github.com/XvoidcrewX/CXA/issues"
Security = "https://github.com/XvoidcrewX/CXA/security"

[project.scripts]
cxa = "python_cli.main:main"
cxa-gui = "python_gui.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["python_*"]
namespaces = false

[tool.setuptools.package-data]
python_core = ["py.typed"]
python_gui = ["*.png", "*.ico", "*.icns"]

[tool.black]
line-length = 100
target-version = ['py310']
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 100
skip = [".venv", "venv"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "E501"]
per-file-ignores = ["__init__.py:F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = false

[tool.bandit]
targets = ["python_core", "python_gui", "python_cli"]
skips = ["B101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]

[tool.coverage.run]
source = ["python_core", "python_gui", "python_cli"]
omit = ["tests/*", "*/__pycache__/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError"
]
